requirements.txt
alembic==1.13.3
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
binwalk==2.3.3
blinker==1.9.0
Brlapi==0.8.3
Brotli==1.0.9
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbfread==2.0.7
dbus-python==1.2.18
decorator==5.2.1
defer==1.0.6
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==3.0.1
Django==5.0.3
django-allauth==65.9.0
django-cors-headers==4.3.1
django-extensions==4.1
django-filter==25.1
django-jazzmin==3.0.1
django-simple-history==3.8.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
dotenv-python==0.0.1
drf-spectacular==0.28.0
duplicity==0.8.21
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.14.1
filelock==3.18.0
Flask==3.1.2
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
flask-shell-ipython==0.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
git-filter-repo==2.47.0
greenlet==3.1.1
gunicorn==23.0.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.1.0
ipdb==0.13.13
ipython==8.34.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.3.5
Markdown==3.8.2
MarkupSafe==3.0.3
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.7
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
packaging==25.0
pandas==2.2.3
paramiko==2.9.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.7
pluggy==1.6.0
ply==3.11
prompt_toolkit==3.0.50
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
Pygments==2.19.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.9.0
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyRFC3339==1.1
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
pytest==8.4.1
pytest-django==4.11.1
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-magic==0.4.24
pythran==0.10.0
pytz==2025.2
pyxdg==0.27
PyYAML==5.4.1
referencing==0.36.2
reportlab==3.6.8
requests==2.32.4
rpds-py==0.27.1
scipy==1.8.0
SecretStorage==3.3.1
sentry-sdk==2.34.1
shiboken6==6.9.2
simplejson==3.17.6
six==1.17.0
soupsieve==2.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.9
systemd-python==234
tkcalendar==1.6.1
tomli==2.2.1
traitlets==5.14.3
ttkbootstrap==1.13.11
typing_extensions==4.13.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.2.0
urllib3==2.5.0
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.29.3
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.10.0
WTForms==3.1.2
xdg==5
xkit==0.0.0
zipp==1.0.0
